requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==14.4.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cohere==5.15.0
colorama==0.4.6
cryptography==45.0.5
dataclasses-json==0.6.7
dnspython==2.7.0
emoji==2.14.1
fastavro==1.11.1
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
greenlet==3.2.3
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.2
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.67
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.4.4
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
olefile==0.47
openai-whisper==20250625
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.3.0
portalocker==2.10.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pypdf==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.3
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.46.1
streamlit-webrtc==0.63.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
unstructured==0.18.2
unstructured-client==0.37.4
urllib3==2.5.0
watchdog==6.0.0
wavio==0.0.9
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
